common ReverseBytes
