+++ can/packer.cc
@@ -8,18 +8,6 @@
 
 #define WARN printf
 
-// this is the same as read_u64_le, but uses uint64_t as in/out
-uint64_t ReverseBytes(uint64_t x) {
-  return ((x & 0xff00000000000000ull) >> 56) |
-          ((x & 0x00ff000000000000ull) >> 40) |
-          ((x & 0x0000ff0000000000ull) >> 24) |
-          ((x & 0x000000ff00000000ull) >> 8) |
-          ((x & 0x00000000ff000000ull) << 8) |
-          ((x & 0x0000000000ff0000ull) << 24) |
-          ((x & 0x000000000000ff00ull) << 40) |
-          ((x & 0x00000000000000ffull) << 56);
-}
-
 static uint64_t set_value(uint64_t ret, const Signal& sig, int64_t ival) {
   int shift = sig.is_little_endian? sig.b1 : sig.bo;
   uint64_t mask = ((1ULL << sig.b2)-1) << shift;
