+++ can/packer_pyx.pyx
@@ -9,7 +9,7 @@ from libcpp cimport bool
 from posix.dlfcn cimport dlopen, dlsym, RTLD_LAZY
 
 from .common cimport CANPacker as cpp_CANPacker
-from .common cimport dbc_lookup, SignalPackValue, DBC
+from .common cimport dbc_lookup, SignalPackValue, DBC, ReverseBytes
 
 
 cdef class CANPacker:
@@ -42,16 +42,6 @@ cdef class CANPacker:
 
     return self.packer.pack(addr, values_thing, counter)
 
-  cdef inline uint64_t ReverseBytes(self, uint64_t x):
-    return (((x & 0xff00000000000000ull) >> 56) |
-           ((x & 0x00ff000000000000ull) >> 40) |
-           ((x & 0x0000ff0000000000ull) >> 24) |
-           ((x & 0x000000ff00000000ull) >> 8) |
-           ((x & 0x00000000ff000000ull) << 8) |
-           ((x & 0x0000000000ff0000ull) << 24) |
-           ((x & 0x000000000000ff00ull) << 40) |
-           ((x & 0x00000000000000ffull) << 56))
-
   cpdef make_can_msg(self, name_or_addr, bus, values, counter=-1):
     cdef int addr, size
     if type(name_or_addr) == int:
@@ -60,5 +50,5 @@ cdef class CANPacker:
     else:
       addr, size = self.name_to_address_and_size[name_or_addr.encode('utf8')]
     cdef uint64_t val = self.pack(addr, values, counter)
-    val = self.ReverseBytes(val)
+    val = ReverseBytes(val)
     return [addr, 0, (<char *>&val)[:size], bus]
