+++ can/parser.py
@@ -1,2 +1,3 @@
 from opendbc.can.parser_pyx import CANParser, CANDefine # pylint: disable=no-name-in-module, import-error
+from opendbc.can.dbc_pyx import DBCParser# pylint: disable=no-name-in-module, import-error
 assert CANParser, CANDefine
